Restore save for local server
